pytest_framework: moving SyslogNgTester to framework
