+++ tests/pytest_framework/functional_tests/destination_drivers/file_destination/test_file_reopen.py
@@ -24,45 +24,7 @@
 from pathlib2 import Path
 import pytest
 
-def with_file_source(variable_name, file_name):
-    def decorator(cls):
-        def add_file_source(self, variable_name, file_name):
-            setattr(self, variable_name, self.config.create_file_source(file_name=file_name))
-        cls.add_init_task(cls, add_file_source, (variable_name, file_name))
-        return cls
-    return decorator
-
-def with_file_destination(variable_name, file_name):
-    def decorator(cls):
-        def add_file_destination(self, variable_name, file_name):
-            setattr(self, variable_name, self.config.create_file_destination(file_name=file_name))
-        cls.add_init_task(cls, add_file_destination, (variable_name, file_name))
-        return cls
-    return decorator
-
-class SyslogNgTester(object):
-    def __init__(self, testcase):
-        self.testcase = testcase
-        self.config = testcase.new_config()
-        self.syslog_ng = testcase.new_syslog_ng()
-        self.syslog_ng_ctl = testcase.new_syslog_ng_ctl()
-
-        for func, args in self.init_functions:
-            func(self, *args)
-
-    def start(self):
-        self.syslog_ng.start(self.config)
-
-    def reopen(self):
-        self.syslog_ng_ctl.reopen()
-
-    @staticmethod
-    def add_init_task(cls, func, args):
-        assert cls != SyslogNgTester
-
-        if not hasattr(cls, "init_functions"):
-            cls.init_functions = []
-        cls.init_functions.append((func, args))
+from src.syslog_ng_tester import *
 
 @with_file_source("file_source", "input.log")
 @with_file_destination("file_destination", "output.log")
